requirements.txt
python-multipart
langchain
pandas
langchain-core
elasticsearch
langchain-community
langchain-neo4j
langchain-elasticsearch
uvicorn
fastapi
vllm
beautifulsoup4
torch==2.5.1
--extra-index-url https://download.pytorch.org/whl/cu121
python-dotenv
sentence-transformers==3.3.1